+++ js/data.js
@@ -1,0 +1,45 @@
+export const description = [
+  'Отель у моря',
+  'Ворота на пляж',
+  'Лазурная бухта',
+  'Девушка фотограф',
+  'Суп с рисом',
+  'Черная машина',
+  'Клубника на тарелке',
+  'Морс в стакане',
+  'Девушка и самолет',
+  'Лоток для обуви',
+  'Пляж',
+  'Белая Ауди',
+  'Салат',
+  'Кот - ролл',
+  'Обувь для дома',
+  'Где-то в небе',
+  'Хор в черном',
+  'Красная машина',
+  'Тапочки с фонариком',
+  'Пальмы в городе',
+  'Макароны в тарелке',
+  'Закат на море',
+  'Мистер Крабс',
+  'На концерте',
+  'Машина и бегемот',
+];
+
+export const message = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export const names = [
+  'Александр',
+  'Ольга',
+  'Дмитрий',
+  'Елена',
+  'Сергей',
+  'Наталья'
+];
